separate out web-related part 1
