+++ daemon/app/GHCSpecter/Render/Web.hs
@@ -1,7 +1,7 @@
 {-# LANGUAGE LambdaCase #-}
 {-# LANGUAGE OverloadedStrings #-}
 
-module GHCSpecter.Render (
+module GHCSpecter.Render.Web (
   render,
 ) where
 
@@ -23,10 +23,10 @@ import GHCSpecter.Channel.Outbound.Types (SessionInfo (..))
 import GHCSpecter.Data.Assets (HasAssets (..))
 import GHCSpecter.Data.Map (forwardLookup, keyMapToList, lookupKey)
 import GHCSpecter.Render.Components.Console qualified as Console
-import GHCSpecter.Render.ModuleGraph qualified as ModuleGraph
-import GHCSpecter.Render.Session qualified as Session
 import GHCSpecter.Render.SourceView qualified as SourceView
-import GHCSpecter.Render.Timing qualified as Timing
+import GHCSpecter.Render.Web.ModuleGraph qualified as ModuleGraph
+import GHCSpecter.Render.Web.Session qualified as Session
+import GHCSpecter.Render.Web.Timing qualified as Timing
 import GHCSpecter.Render.Util (divClass)
 import GHCSpecter.Server.Types (
   HasServerState (..),
@@ -182,3 +182,4 @@ render (ui, ss) =
   case ui ^. uiModel . modelTransientBanner of
     Just fraction -> renderBanner (ui ^. uiAssets . assetsGhcSpecterPng) fraction
     Nothing -> renderMainView (ui ^. uiModel, ss)
+
