+++ daemon/ghc-specter-daemon.cabal
@@ -32,7 +32,6 @@ library
       GHCSpecter.Data.GHC.Hie
       GHCSpecter.Data.Timing.Types
       GHCSpecter.Data.Timing.Util
-      GHCSpecter.Driver
       GHCSpecter.Driver.Comm
       GHCSpecter.Driver.Session
       GHCSpecter.Driver.Session.Types
@@ -43,7 +42,6 @@ library
       GHCSpecter.GraphLayout.OGDF
       GHCSpecter.GraphLayout.Sugiyama
       GHCSpecter.GraphLayout.Types
-      GHCSpecter.Render
       GHCSpecter.Render.Components.Console
       GHCSpecter.Render.Components.ConsoleItem
       GHCSpecter.Render.Components.GHCCore
@@ -53,11 +51,9 @@ library
       GHCSpecter.Render.Components.TextView
       GHCSpecter.Render.Components.TimingView
       GHCSpecter.Render.ConcurReplicaSVG
-      GHCSpecter.Render.ModuleGraph
       GHCSpecter.Render.Session
       GHCSpecter.Render.SourceView
       GHCSpecter.Render.Tab
-      GHCSpecter.Render.Timing
       GHCSpecter.Render.Util
       GHCSpecter.Server.Types
       GHCSpecter.UI.ConcurReplica.DOM
@@ -137,6 +133,11 @@ library
 executable ghc-specter-daemon
   main-is: Main.hs
   other-modules:
+      GHCSpecter.Driver.Web
+      GHCSpecter.Render.Web
+      GHCSpecter.Render.Web.ModuleGraph
+      GHCSpecter.Render.Web.Session
+      GHCSpecter.Render.Web.Timing            
       Paths_ghc_specter_daemon
   hs-source-dirs:
       app
@@ -145,11 +146,20 @@ executable ghc-specter-daemon
       aeson
     , base >=4.15 && <5
     , bytestring
+    , concur-core
+    , concur-replica
+    , containers
+    , errors
+    , extra
     , ghc-specter-daemon
     , ghc-specter-plugin
+    , ghc-specter-render
+    , lens
     , optparse-applicative
+    , pretty-simple
     , stm
     , text
+    , time
   default-language: GHC2021
 
 test-suite ghc-specter-daemon-test
